Refactor: group functions by loaded DLL modules
